Initial commit for the refactoring of the _discrete module
